[Feature] added BitOffset's, these are syntantic sugor with checks, like a = b[10:6];
BitOffsets are used to copy certain bits of an unsigned int32/int64. This is
often done in drivers like:

(ANSI-C)
int bitfield = ((regval >> 10) & 0x1F);

With C2's BitOffsets, this can be written as:

int bitfield = regval[14:10];

Additionally, possibleoverflows can be checked (by bitwidth)
NOTE that this is something else than BitFields, which are ways to pack x-bit values
in a struct.