NEW Feature! - struct-functions
Struct-functions allow C2 code to be more readable, speeding up development.
In a nutshell, they work as follows.
Without struct-functions:
   point_add(&p1, 10);
With struct-functions:
    p1.add(10);

The rest of the code doesn't change! (so both p1 and point_add() are defined
as usual)

For info see the website:
http://www.c2lang.org/site/language/struct_functions
or check the tests: tests/Functions/struct_functions/