added Vulkan example + lib
The example is a C2 port of https://github.com/SaschaWillems/Vulkan,
the triangle example. It's not complete yet, since it uses a C++
Matrix lib that needs to have a C/C2 variant first, but it does show
a (glfw) window and handles events.

Porting the Vulkan interface itself was quite straightforward and
almost nothing more than a mechanical process.